feat: logged in session for all tests
