--- frontend/src/__e2e__/Dashboard.spec.ts
@@ -1,7 +1,0 @@
-import { expect } from '@playwright/test';
-import test from './config/AuthTest';
-
-test('Dashboard', async ({ page }) => {
-  await page.goto('/');
-  await expect(page).toHaveTitle(/Silva/);
-});
