+++ frontend/src/__e2e__/auth/auth.setup.ts
@@ -1,21 +1,34 @@
-import { test, expect } from '@playwright/test';
+import { chromium, expect } from '@playwright/test';
+import path from 'path';
+import { fileURLToPath } from 'url';
 
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const authFile = path.join(__dirname, './user.json');
 const baseURL = process.env.BASE_URL ?? 'http://localhost:3000';
 const bceidUser = process.env.TEST_BCEID_USERNAME ?? '';
 const bceidPassword = process.env.TEST_BCEID_PASSWORD ?? '';
 
-test('visualize login flow', async ({ page }) => {
+
+async function globalSetup() {
+  const browser = await chromium.launch();
+  const page = await browser.newPage();
+
   await page.goto(baseURL);
   await page.click('[data-testid="landing-button__bceid"]');
 
   await page.waitForSelector('#user');
   await page.fill('#user', bceidUser);
-
   await page.fill('#password', bceidPassword);
-
   await page.click('input[name="btnSubmit"]');
 
   await page.waitForURL('**/dashboard');
-
   await expect(page.getByRole('heading', { name: 'Dashboard' })).toBeVisible();
-});
+
+  await page.context().storageState({ path: authFile });
+  await browser.close();
+};
+
+
+export default globalSetup;
