+++ frontend/src/__e2e__/auth/user.json
@@ -1,0 +1,1 @@
+{}
